% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ad.R
\name{add_ad}
\alias{add_ad}
\title{add_ad}
\usage{
add_ad(
  newad,
  newad_cat = NULL,
  newad_subcat = NULL,
  newad_customer,
  newad_click_rate
)
}
\arguments{
\item{newad}{new advertisement. Character string of name of new advertisement.}

\item{newad_cat}{new advertisement category. Character string of name of new advertisement category. If unavailable a random category from those already available in ads will be used.}

\item{newad_subcat}{new advertisement subcategory. Character string of name of new advertisement subcategory. If unavailable a random subcategory from those already available in catsToSubcats and associated with the given category will be used.}

\item{newad_customer}{new advertisement category. Integer indicating the customer id.}

\item{newad_click_rate}{new advertisement rate.}
}
\value{
dataframe
}
\description{
Add a new advertisement to the dataframe of advertisements.
The function assumes the existence of a file "ads_file":
 - a dataframe ads, with minimal structure "ads" "subcats" "cats" "customer";
}
