% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{onLoad hook}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{A character string giving the library directory where the package defining the namespace was found.}

\item{pkgname}{A character string giving the name of the package.}
}
\value{
Side effects only. Defines example data and support functions and puts them in the global environment.
}
\description{
onLoad hook
}
