% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.ads}
\alias{read.ads}
\alias{write.ads}
\alias{recreate.ads}
\alias{read.users}
\alias{write.users}
\alias{recreate.users}
\title{Utils to read, write and restore from backup example ads and users data}
\usage{
read.ads(...)

write.ads(x, ...)

recreate.ads(...)

read.users(...)

write.users(object, ...)

recreate.users(...)
}
\arguments{
\item{...}{passed down to the low level read/write functions}

\item{x}{the object to be written by write.csv2}

\item{object}{the object to be written by saveRDS}
}
\description{
convenient wrappers to read.csv2 and write.csv2 frequent calls in adspool.
If ever this package is developed to read-write from database, it would be sufficient to update these wrappers.
}
\details{
Read advertisements from disk
}
\section{Functions}{
\itemize{
\item \code{write.ads}: Write advertisements from disk

\item \code{recreate.ads}: Re-create the ads file

\item \code{read.users}: Read users from disk

\item \code{write.users}: Write users from disk

\item \code{recreate.users}: Re-create the users file
}}

