% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ads.R
\name{remove_ads}
\alias{remove_ads}
\title{remove_ads}
\usage{
remove_ads(
  ad_id = NULL,
  ad_name = NULL,
  ad_cat = NULL,
  ad_subcat = NULL,
  ad_client = NULL
)
}
\arguments{
\item{ad_id}{advertisement id. Integer positive number.}

\item{ad_name}{advertisement name. Character vector with the name of the advertisement}

\item{ad_cat}{category of the advertisements to return. Character vector.}

\item{ad_subcat}{subcategory of the advertisements to return. Character vector.}

\item{ad_client}{client of the advertisements to return. Character vector.}
}
\value{
dataframe
}
\description{
Remove entry(-es) from the dataframe of advertisements via filtering.
}
