% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_user_info.R
\name{return_user_info}
\alias{return_user_info}
\title{return_user_info}
\usage{
return_user_info(id, ...)
}
\arguments{
\item{id}{vector of integers representing users}

\item{...}{placeholder}
}
\value{
dataframe
}
\description{
Return infos for given user
}
