% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ads.R
\name{select_ads}
\alias{select_ads}
\title{select_ad}
\usage{
select_ads(n = 1L, onlyMatching = FALSE, ...)
}
\arguments{
\item{n}{Number of adverts to return}

\item{onlyMatching}{Logical. When \code{TRUE} return only ads matching filter criteria}

\item{...}{Additional parameters for the ad selection, including ads filter criteria
to be passed on to \link{subset_ads}}
}
\value{
\code{data.frame} with metadata of the advertisements
}
\description{
Select the best adverts to be exposed in the given input context
}
\details{
When \code{onlyMatching = TRUE} and there are no matching adverts in the pool, then the result is empty.
Otherwise, the resulting data frame always contains \code{n} elements. When possible, the adverts are
taken from the pool of ads with matching criteria. Matching ads are prioritized over non matching ones.
If \code{onlyMatching = TRUE} and the number of matching ads in the pool is less than \code{n}, then these will
be recycled until filling the request.
}
