% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_ads.R
\name{subset_ads}
\alias{subset_ads}
\title{subset_ads}
\usage{
subset_ads(
  ad_id = NULL,
  ad_name = NULL,
  ad_cat = NULL,
  ad_subcat = NULL,
  ad_customer = NULL
)
}
\arguments{
\item{ad_id}{advertisement id. Integer positive number.}

\item{ad_name}{advertisement name. String with the name of the advertisement}

\item{ad_cat}{category of the advertisements to return. String.}

\item{ad_subcat}{subcategory of the advertisements to return. String.}

\item{ad_customer}{customer of the advertisements to return. String.}
}
\value{
\code{dataframe}, a subset of the ads that match the given parameters.
\code{ad_name}, \code{ad_cat} and \code{ad_subcat} are matched by pattern, \code{ad_id} and \code{customer_id} exact match.
}
\description{
Return subset of the dataframe of advertisements via filtering.
The function assumes the existence of a file "ads_file":
 - a dataframe ads, with minimal structure "ads" "subcats" "cats" "customer";
}
