% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ad.R
\name{update_ad}
\alias{update_ad}
\title{update_ad}
\usage{
update_ad(
  ad_id,
  ad_name = NULL,
  ad_cat = NULL,
  ad_subcat = NULL,
  ad_customer = NULL,
  ad_click_count = NULL,
  ad_click_rate = NULL
)
}
\arguments{
\item{ad_id}{advertisement id. Integer positive number.}

\item{ad_name}{advertisement name. Character vector with the name of the advertisement}

\item{ad_cat}{category of the advertisements to return. Character vector.}

\item{ad_subcat}{subcategory of the advertisements to return. Character vector.}

\item{ad_customer}{customer of the advertisements to return. Character vector.}

\item{ad_click_count}{click count}

\item{ad_click_rate}{price per click}
}
\value{
dataframe
}
\description{
Update an entry from the dataframe of advertisements.
}
