% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_history_ads.R
\name{update_history_ads}
\alias{update_history_ads}
\title{update_history_ads}
\usage{
update_history_ads(ads.history)
}
\arguments{
\item{ads.history}{List containing the history of ads of a particular session.}
}
\value{
Side effect only.
}
\description{
Update the history of the ads shown.
}
