% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write.logs}
\alias{write.logs}
\title{Append info to logging file}
\usage{
write.logs(line, ...)
}
\arguments{
\item{line}{the object to be written by write}

\item{...}{passed down to the low level functions}
}
\description{
convenient wrappers to write() for frequent calls in adspool.
}
\details{
Write logging info to disk
}
